Relocate CLI related packages
