+++ internal/command/init/common_test.go
