+++ internal/command/run/run.go
@@ -8,6 +8,7 @@ import (
 	"github.com/saucelabs/saucectl/internal/playwright"
 	"github.com/saucelabs/saucectl/internal/puppeteer"
 	"github.com/saucelabs/saucectl/internal/testcafe"
+	version2 "github.com/saucelabs/saucectl/internal/version"
 	"github.com/saucelabs/saucectl/internal/xcuitest"
 	"net/http"
 	"os"
@@ -20,7 +21,6 @@ import (
 	"github.com/rs/zerolog/log"
 	"github.com/spf13/cobra"
 
-	"github.com/saucelabs/saucectl/cli/version"
 	"github.com/saucelabs/saucectl/internal/appstore"
 	"github.com/saucelabs/saucectl/internal/config"
 	"github.com/saucelabs/saucectl/internal/credentials"
@@ -145,7 +145,7 @@ func Command() *cobra.Command {
 
 // preRun is a pre-run step that is executed before the main 'run` step. All shared dependencies are initialized here.
 func preRun() error {
-	println("Running version", version.Version)
+	println("Running version", version2.Version)
 	checkForUpdates()
 	go awaitGlobalTimeout()
