+++ internal/flags/emulator.go
