+++ internal/sentry/sentry.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"github.com/getsentry/sentry-go"
 	"github.com/rs/zerolog/log"
-	"github.com/saucelabs/saucectl/cli/setup"
+	setup2 "github.com/saucelabs/saucectl/internal/setup"
 	"io"
 	"mime/multipart"
 	"net/http"
@@ -60,7 +60,7 @@ func attach(client http.Client, eventID, filename string) {
 	_, _ = io.Copy(fw, r)
 	_ = w.Close()
 
-	sentryURL, err := attachmentURLFromDSN(setup.SentryDSN, eventID)
+	sentryURL, err := attachmentURLFromDSN(setup2.SentryDSN, eventID)
 	if err != nil {
 		log.Debug().Err(err).Msg("Failed to assemble the sentry attachment URL")
 		return
